refactor: Grouped decideVersion under utilities for blueprints
