+++ src/utils/blueprints.js
@@ -1,1 +1,2 @@
+export * from './blueprints/decide-version.js';
 export * from './blueprints/process-template.js';
